more value-classes
