+++ core/src/main/scala/org/hammerlab/iterator/start/HeadOption.scala
@@ -1,0 +1,12 @@
+package org.hammerlab.iterator.start
+
+import hammerlab.iterator.macros.IteratorOps
+
+@IteratorOps
+class HeadOption[T](it: BufferedIterator[T]) {
+  def headOption: Option[T] =
+    if (it.hasNext)
+      Some(it.head)
+    else
+      None
+}
