+++ core/src/main/scala/org/hammerlab/iterator/start/NextOption.scala
@@ -1,0 +1,12 @@
+package org.hammerlab.iterator.start
+
+import hammerlab.iterator.macros.IteratorOps
+
+@IteratorOps
+class NextOption[T](it: Iterator[T]) {
+  def nextOption: Option[T] =
+    if (it.hasNext)
+      Some(it.next)
+    else
+      None
+}
